Trigger transition during save instead of response_change
Move the fsm transition trigger to save_model so we can take
advantage of the built in logging that already happens when a
save is processed.  This also removes a bug which caused a
double log entry would be created, one for the state transition
and a 'No fields changed' entry that represented a lack of changes
to other fields.

In addition, this also causes the transition to correctly create
a revision if django-reversion is being used.