Add support for multiple fsm fields
Sometimes we want to be able to transition more than one fsm state
field, so add support for a list of fields to be specified.  Attempt
to retain some backward compatibility by retrieving state fields via
a method that will attempt to "fix" a single field by converting it
to a list.

Convert the inline creation of the state transition buttons to an
inclusion tag in order to hide the implementation of the transitions
in the context from the submit line template.  Include the field
name as part of the button name so we can properly negotiate our
transition.